move sat solving to separate file
